job module pr changes

+++ src/job/job.resolver.ts
@@ -13,6 +13,7 @@ import { CreateJobInput } from './dto/create-job.input';
 import { UpdateJobInput } from './dto/update-job.input';
 import { Job } from './job.model';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { BadRequestException } from '@nestjs/common';
 
 @Resolver(() => Job)
 export class JobResolver {
@@ -23,6 +24,15 @@ export class JobResolver {
 
   @Mutation(() => Job)
   async createJob(@Args('data') createJobInput: CreateJobInput) {
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id: createJobInput.profileId,
+      },
+      select: { type: true },
+    });
+    if (!check) {
+      return new BadRequestException({ error: 'Recruiter can only post jobs' });
+    }
     return await this.jobService.create(createJobInput);
   }
 
@@ -41,12 +51,34 @@ export class JobResolver {
     @Args('id', { type: () => ID }) id: string,
     @Args('data') updateJobInput: UpdateJobInput,
   ) {
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id: updateJobInput.profileId,
+      },
+      select: { type: true },
+    });
+    if (!check) {
+      return new BadRequestException({
+        error: 'Recruiter can only update jobs',
+      });
+    }
     await this.jobService.findOne(id);
     return this.jobService.update(id, updateJobInput);
   }
 
   @Mutation(() => Job)
   async removeJob(@Args('id', { type: () => ID }) id: string) {
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id,
+      },
+      select: { type: true },
+    });
+    if (!check) {
+      return new BadRequestException({
+        error: 'Recruiter can only delete jobs',
+      });
+    }
     await this.jobService.findOne(id);
     return this.jobService.remove(id);
   }
